requirement.txt
asgiref==3.8.1
certifi==2025.4.26
channels==4.3.1
channels-redis
charset-normalizer==3.4.2
cloudinary==1.44.0
daphne==4.2.1
Django==5.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-extensions
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
redis
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0